Use CC header to publish mt msgs
The greate deal of code that processed pair of publish
confirms id (the first for kelly queue and the second
for the just gtw queue) was replaced with via rabbitMQ
extention - CC header. This solution allowed to publish
msgs with single publish function call and also with
single publish confirm callback.